server/requirements.txt
asgiref>=3.7.0
boto3>=1.34.0
cffi>=1.16.0
cryptography>=42.0.0
DateTime>=5.4
Django>=4.2.0,<5.0
django-cors-headers>=4.3.0
django-environ>=0.11.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
earthengine-api>=0.1.400
pillow>=10.2.0
psycopg2-binary>=2.9.0
pycparser>=2.21
PyJWT>=2.8.0
python-dotenv>=1.0.0
pytz>=2024.1
sqlparse>=0.4.0
zope.interface>=6.0
